remove useless parts and rename the plugin
